Refactor app shell routing and extract global home
